fix(discovery): move claudeReAsked to lib to satisfy Next.js route exports
Next.js route files only allow HTTP-verb exports.  Extract keyword
detection into src/lib/ai/discovery-progress.ts and import from there.
